Refactor wrong user warning to be a helper
